+++ packages/qora/lib/qora.dart
@@ -1,6 +1,6 @@
 // packages/reqry/lib/reqry.dart
 
-/// Reqry - Re-query with ease
+/// Qora
 ///
 /// A powerful async state management library for Dart,
 /// inspired by TanStack Query.
@@ -8,17 +8,15 @@ library;
 
 // Core
 export 'src/core/cached_entry.dart';
+export 'src/core/key/qora_key.dart';
 export 'src/core/persist_qora_client.dart';
 export 'src/core/qora_client.dart';
 export 'src/core/qora_client_config.dart';
-export 'src/core/qora_key.dart';
 export 'src/core/qora_options.dart';
 export 'src/core/qora_state.dart';
 export 'src/core/query_function.dart';
-
 // Managers
 export 'src/managers/connectivity_manager.dart';
 export 'src/managers/lifecycle_manager.dart';
-
 // Utils
 export 'src/utils/qora_exception.dart';